Create main.py
There are some spherical balloons spread in two-dimensional space. For each balloon, provided input is the start and end coordinates of the horizontal diameter. Since it's horizontal, y-coordinates don't matter, and hence the x-coordinates of start and end of the diameter suffice. The start is always smaller than the end.

An arrow can be shot up exactly vertically from different points along the x-axis. A balloon with xstart and xend bursts by an arrow shot at x if xstart ≤ x ≤ xend. There is no limit to the number of arrows that can be shot. An arrow once shot keeps traveling up infinitely.

Given an array points where points[i] = [xstart, xend], return the minimum number of arrows that must be shot to burst all balloons.

Explanation of code:
1. Sorting by the points by point[i][1]; where i = len(points)
2. Comparing current pointer with point[i][0]; where i = len(points) 
3. Assigning the current pointer with the new points[i]; where i = len(points) if Point#2 is correct

Time complexity: O(n)